Training Ground Setup - Step 3
Created a small tile palette to start basic design of game structure for the Player to run and attack Enemies. Installed Cinemachine plugin from Unity Store and use a virtual camera to follow the Player. Setup a parallax background so that the sky and building backgrounds move at a different rate of speed which is a nice effect for gameplay.