Enemy Finite State Machine - Step 4
Very similar to the player's state machine, I made a slightly scaled down version of the enemy state machine to handle idle, moving and attacking animations. Created an Entity script so that the Player and Enemy characters can share inheritance for similar variables and functions. Next commit will focus on a basic battle system so the Player and the Enemy can collide with one another and (eventually) cause damage to each other.