Sword Skill - Step 8
Created three states for the sword skill that handle throwing, aiming and catching the sword. Also created four different kinds of swords, all of which react different to enemies when thrown including regular, bounce, pierce and spin swords. Added a behavior that temporary freezes the animation of the Enemy when hit with one of the swords. Finally during this step, I renamed the Entity class to the Character class to better reflect the nature of the class. A ton of coding done for this step. In the next commit, we will work on the Ultimate skill of the Player.